+++ document/RELEASES.md
@@ -43,3 +43,11 @@ Based on:
 - Speakeasy CLI 1.222.0 (2.287.0) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.3.0] document
+
+## 2024-03-27 00:12:25
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/document.yaml
+- Speakeasy CLI 1.227.0 (2.291.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.3.1] document